made changes to the "detection methodology"
based on constructive feedback from a tester.

responses with a content-type that is JSON-ish
no longer need to satisfy pathname restrictions.
they are, however, still inspected for a JSONP signature.

added the hash '#JSON-DataView' as another way
to signal to the addon that the response is JSON;
for use when the content-type is weak,
and the pathname fails to include a proper file extension.