reverted 3rd-party libraries back to unmodified original versions
the Firefox AMO reviewers can't seem to make up their minds
as to what they want.

1st, the addon was rejected because of certain methods being present
in 3rd-party libraries, and I was required to sanitize them.

next, the addon was rejected because the hash/checksum
of the 3rd-party libraries doesn't match the original.

I'm honestly at a loss at this point..

3rd party libraries reverted to:
* https://github.com/beautify-web/js-beautify/raw/04645a19b2a3ed05cff177644016ee89c5343d19/js/lib/beautify.js