[e10s/v2.0] support for multi-process (Electrolysis, e10s)
Admittedly, weak support.

Pretty much all of the code needs access to the `document` element.
1st for inspection, to see whether criteria are met for making changes.
2nd for applying changes to.

Rather than migrating the entire add-on into `frame` script(s),
a trivial document-ready event listener was moved from `chrome` to `frame`.
When this event fires (in `frame`), it has access to the `document` element.
An async message is then sent from `frame` to `chrome`,
which passes a reference to the `document` element.

Question:
=========
How inefficient is this methodology?
The document-ready event listener fires on every page load;
resulting in the `document` element being passed inter-process to `chrome`.
For non-JSON pages, the code that runs in `chrome` will inspect the `document`
and quickly determine that no work needs to be done.
But this overall strategy feels like cheating,
and I'm worried about performance implications.